pyproject.toml
[project]
name = "autobook"
version = "0.1.0"
description = "Telegram bot for car seat reservation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.0",
    "aiosqlite",
    "python-dotenv",
]
